+++ classes/Order.h
@@ -33,7 +33,7 @@ class HashTable;
 
 class Order {
 public:
-	uint32_t		order_id;
+	uint64_t		order_id;
 	uint32_t		account_id;
 	Market			*market;
 	order_type_t	direction;
@@ -44,7 +44,7 @@ public:
 
 	std::vector<Transaction *>	transactions;
 
-	Order(Market *market, const uint32_t account_id, const order_type_t direction, const uint64_t qty, const uint64_t price, const uint32_t order_id, const uint32_t timestamp, const order_status_t status);
+	Order(Market *market, const uint32_t account_id, const order_type_t direction, const uint64_t qty, const uint64_t price, const uint64_t order_id, const uint32_t timestamp, const order_status_t status);
 	Order(Market *market, const uint32_t account_id, const order_type_t direction, const uint64_t qty, const uint64_t price);
 	virtual ~Order();
